Made package name singular
