+++ src/main/java/sc/fiji/bdv/source/read/SourceAdder.java
@@ -1,12 +1,10 @@
-package sc.fiji.bdv.sources.read;
+package sc.fiji.bdv.source.read;
 
 import bdv.util.BdvFunctions;
 import bdv.util.BdvHandle;
 import bdv.util.BdvOptions;
 import bdv.viewer.Source;
-import sc.fiji.bdv.BdvUtils;
 import sc.fiji.bdv.navigate.ViewerTransformAdjuster;
-import sc.fiji.bdv.sources.display.BrightnessAdjuster;
 
 public class SourceAdder implements Runnable
 {
